Reorganizing nonlinear solvers that now use JAX
